requirements.txt
cryptography>=41.0.0
asyncpg>=0.29.0
redis>=5.0.0
click>=8.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pqcrypto-dilithium>=0.1.0
secretsharing>=0.2.0
qrcode>=7.4.0
Pillow>=10.0.0
